+++ lib/features/auth/domain/repositories/auth_repository.dart
@@ -3,4 +3,6 @@ import 'package:mobile/features/auth/domain/entities/user_entity.dart';
 abstract class AuthRepository {
   Future<UserEntity> signInWithGoogle();
   Future<bool> requestDriveScope();
+  Future<String?> getDriveAccessToken();
+  Future<void> signOutFromGoogle();
 }