+++ lib/features/auth/domain/use_cases/sign_out_from_google_usecase.dart
@@ -1,0 +1,11 @@
+import 'package:mobile/features/auth/domain/repositories/auth_repository.dart';
+
+class SignOutFromGoogleUseCase {
+  final AuthRepository repository;
+
+  SignOutFromGoogleUseCase(this.repository);
+
+  Future<void> call() async {
+    await repository.signOutFromGoogle();
+  }
+}